classification/requirements.txt
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
attrs==22.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1677499911949/work
brotlipy==0.7.0
certifi==2023.5.7
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
chardet @ file:///home/builder/ci_310/chardet_1640804867535/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
conda @ file:///home/conda/feedstock_root/build_artifacts/conda_1685035123102/work
conda-build==3.23.3
conda-package-handling @ file:///croot/conda-package-handling_1666940373510/work
cryptography @ file:///croot/cryptography_1665612644927/work
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dnspython==2.2.1
exceptiongroup==1.0.4
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
expecttest==0.1.4
filelock @ file:///opt/conda/conda-bld/filelock_1647002191454/work
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
future==0.18.2
fvcore @ file:///home/conda/feedstock_root/build_artifacts/fvcore_1671623667463/work
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
hypothesis==6.61.0
idna @ file:///croot/idna_1666125576474/work
iopath @ file:///home/conda/feedstock_root/build_artifacts/iopath_1636568816351/work
ipython @ file:///croot/ipython_1670919316550/work
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1632759296524/work
jsonpointer==2.0
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
MarkupSafe @ file:///home/builder/ci_310/markupsafe_1640791577099/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
mpmath==1.2.1
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.3.0
pkginfo @ file:///croot/pkginfo_1666725041340/work
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
portalocker @ file:///home/conda/feedstock_root/build_artifacts/portalocker_1674135655894/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-etcd==0.4.5
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sortedcontainers==2.4.0
soupsieve @ file:///croot/soupsieve_1666296392845/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.11.1
tabulate @ file:///home/conda/feedstock_root/build_artifacts/tabulate_1665138452165/work
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1682317048417/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz @ file:///croot/toolz_1667464077321/work
torch==1.13.1
torchelastic==0.2.2
torchtext==0.14.1
torchvision==0.14.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets @ file:///croot/traitlets_1671143879854/work
types-dataclasses==0.6.6
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3 @ file:///croot/urllib3_1670526988650/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
yacs @ file:///home/conda/feedstock_root/build_artifacts/yacs_1645705974477/work
